requirements.txt
cupy_cuda113==10.6.0
matplotlib==3.7.3
numpy==1.22.4
pandas==2.0.3
scikit_learn==1.0.2
scipy==1.7.3
sympy==1.11.1
torch==1.12.1
vmdpy==0.2
